Build: reuse or start volumes for CI run

+++ .teamcity/run_ci.sh
@@ -165,8 +165,13 @@ function docker_pull() {
     fi
 }
 
-function docker_start_volume() {
+# Note(storypku): Reuse existing docker volumes for CI
+function reuse_or_start_volume() {
     local container="$1"
+    if docker ps --format "{{.Names}}" | grep -q "${container}" ; then
+        info "Found existing volume \"${container}\", will be reused."
+        return
+    fi
     local image="$2"
     docker_pull "${image}"
     docker run -id --rm --name "${container}" "${image}"
@@ -186,7 +191,7 @@ function start_map_volume() {
             map_image="${DOCKER_REPO}:map_volume-${map_name}-${map_version}"
         fi
         info "Load map ${map_name} from image: ${map_image}"
-        docker_start_volume "${map_volume}" "${map_image}"
+        reuse_or_start_volume "${map_volume}" "${map_image}"
         MAP_VOLUME_CONF="${MAP_VOLUME_CONF} --volumes-from ${map_volume}"
     fi
 }
@@ -195,7 +200,7 @@ function mount_map_volumes() {
     info "Starting mounting map volumes ..."
     if [ "$FAST_MODE" = "no" ]; then
         for map_name in ${DEFAULT_MAPS[@]}; do
-            start_map_volume ${map_name} "${VOLUME_VERSION}"
+            start_map_volume "${map_name}" "${VOLUME_VERSION}"
         done
     else
         for map_name in ${DEFAULT_TEST_MAPS[@]}; do
@@ -211,20 +216,20 @@ function mount_other_volumes() {
         # YOLO3D
         local yolo3d_volume="apollo_yolo3d_volume_${USER}"
         local yolo3d_image="${DOCKER_REPO}:yolo3d_volume-${TARGET_ARCH}-latest"
-        docker_start_volume "${yolo3d_volume}" "${yolo3d_image}"
+        reuse_or_start_volume "${yolo3d_volume}" "${yolo3d_image}"
         volume_conf="${volume_conf} --volumes-from ${yolo3d_volume}"
     fi
 
     # LOCALIZATION
     local localization_volume="apollo_localization_volume_${USER}"
     local localization_image="${DOCKER_REPO}:localization_volume-${TARGET_ARCH}-latest"
-    docker_start_volume "${localization_volume}" "${localization_image}"
+    reuse_or_start_volume "${localization_volume}" "${localization_image}"
     volume_conf="${volume_conf} --volumes-from ${localization_volume}"
 
     if [ "${TARGET_ARCH}" = "x86_64" ]; then
         local local_3rdparty_volume="apollo_local_third_party_volume_${USER}"
         local local_3rdparty_image="${DOCKER_REPO}:local_third_party_volume-${TARGET_ARCH}-latest"
-        docker_start_volume "${local_3rdparty_volume}" "${local_3rdparty_image}"
+        reuse_or_start_volume "${local_3rdparty_volume}" "${local_3rdparty_image}"
         volume_conf="${volume_conf} --volumes-from ${local_3rdparty_volume}"
     fi
